EE10.06: Extract cli-runner.ts; reduce orchestrator.ts to pure barrel; realign test imports
- Move runDeliveryOrchestrator and private helpers into cli-runner.ts
- Reduce orchestrator.ts to pure re-export barrel (no logic, no state)
- Delete deprecated recordInternalReview
- Migrate EE8.01/EE8.02 describe blocks from orchestrator.test.ts to ticket-flow.test.ts
- Update orchestrator.test.ts to import from source modules, not the barrel
